ner/requirements.txt
PyYAML==6.0.2
torch==1.13.1
numpy==1.26.4
einops==0.8.0
pandas==2.2.3
plotly==5.24.1
gradio==5.12.0
torchvision==0.14.1
